ASoC: soc-pcm: fix a codec fixup issue in TDM case
On HDaudio platforms, if playback is started when capture is working,
there is no sound produced by playback. In ASoC, stream tag for
HDaudio is stored in rx|tx_mask for codec. If capture is stared
before playback, rx_mask would be non-zero on HDaudio platform,
and params fixup function based on rx_mask would change the channel
number of playback in hw_params for codec.

According to the author of tx|rx_mask, tx_mask is for playback
and rx_mask is for capture. And stream direction is checked at all
other references of tx|rx_mask in ASoC, so here should be an error.

This issue would affect not only HDaudio+ASoC, but also I2S codecs if
the channel number based on rx_mask is not equal to the one for tx_mask.
It could be rarely reproduecd because most drivers in kernel set the same
channel number to tx|rx_mask or rx_mask is zero.

Tested on all platforms using stream_tag & HDaudio and intel I2S platforms.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>