ASoC: intel: skl: fix a simultaneous playback & capture issue on hda platform
If playback and capture are working simultaneously and
capture is stopped, there is no sound produced by playback.
snd_hdac_ext_link_clear|set_stream_id are for playback to map
or unmap stream id with link output, so capture should not call
them.

This patch add check for stream direction before calling this
function. I don't do it in these stream functions for lacking of
stream direction

Test on Apollolake and Kabylake with SST driver.

Signed-off-by: Rander Wang <rander.wang@linux.intel.com>